requirements.txt
aiohttp==3.6.3
astroid==2.4.2
async-timeout==3.0.1
attrs==20.2.0
chardet==3.0.4
colorama==0.4.4
discord.py==1.5.1
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.1
regex==2020.10.28
six==1.15.0
termcolor==1.1.0
toml==0.10.2
tzlocal==2.1
wrapt==1.12.1
yarl==1.5.1
